us_fertilizer_county
Source: R/data.R

This is data adapted from the county-level estimates of fertilizer nitrogen and phosphorus based on commercial sales from 1945 to 2012. Please visit [here](https://www.sciencebase.gov/catalog/item/5851b2d1e4b0f99207c4f238) for more details.

us_fertilizer_county

Format

A data frame with 582012rows and 11 variables:

FIPS

FIPS is a combination of state and county codes, in character format.

State

The state abbr. of U.S.

County

County name in U.S.

ALAND

The land area in each county, unit: km squared

AWATER

The water area in each county, unit: km squared

INTPTLAT

The latitude of centriod in each county, e.g. 32.53638

INTPTLONG

The longitude of centriod in each county, e.g. -86.64449

Quantity

The quantity of fertilizeation as N or P, e.g. kg N or kg P

Year

The year of estimated data, e.g. 1994

Nutrient

The fertilizer type, e.g. N or P

Farm.Type

The land use type of fertilizer, e.g. farm and nonfarm

Input.Type

The input type of nutrient, e.g. Fertilizer or Manure

...

Examples

  require(usfertilizer)
  data(us_fertilizer_county)